SINGA-36 Clean ModelProto, ClusterProto, JobProto and driver program
The ClusterProto and ModelProto are merged into JobProto.
A SingaProto is added for global singa configuration including zookeeper info, log dir setting.

The driver program (i.e., main.cc) cleaned.
1. Users include singa.h and their own headers the the drvier program.
2. They prepare the job configuration
and singa configuration (may make this transparent to users, e.g., through environment variable, SINGA_HOME).
3. They register the layers them implemented.
4. Call the trainer to start the training
After linking the driver program with libsinga, users pass the executable to bin/singa-run.sh
together with the workspace where job.conf is stored to submit the job

Some layers are moved into optional_layer.cc(.h) if they have dependencies on external libs (e.g., lmdb and opencv).
If users do not need these layers, they do not need to install the dependencies. TODO update configure script to add
these options. USE_OPTIONAL_LAYER controls the compliation of optional_layer.cc(.h).

Move command line argument singa_conf into trainer.cc.
Usually, users do not need to pass this argument explicitly.
The sing-run.sh script will find it automatically.

modify scripts to support new singa command line

minor change on example job config

add SubmitJob func in singa.h to simplify the job submission API.
