+++ conf/singa.conf
@@ -2,4 +2,4 @@
 zookeeper_host: "localhost:2181"
 
 # set if you want to change log directory
-# log_dir: "/tmp/singa-log/"
+log_dir: "/tmp/singa-log/"
