--- examples/mnist/cluster.conf
@@ -1,7 +1,0 @@
-nworker_groups: 1
-nserver_groups: 1
-nservers_per_group: 1
-nworkers_per_group: 1
-nservers_per_procs: 1
-nworkers_per_procs: 1
-workspace: "examples/mnist/"
