+++ examples/mnist/conv.conf
@@ -1,3 +1,13 @@
+cluster {
+  nworker_groups: 1
+  nserver_groups: 1
+  nservers_per_group: 1
+  nworkers_per_group: 1
+  nservers_per_procs: 1
+  nworkers_per_procs: 1
+  workspace: "examples/mnist/"
+}
+model {
 name: "mnist-conv"
 train_steps: 10000
 test_steps:100
@@ -171,5 +181,6 @@ layer{
   }
   srclayers:"ip2"
   srclayers:"label"
+}
 }
 }
