+++ src/main.cc
@@ -1,29 +1,19 @@
-#include <gflags/gflags.h>
-#include <glog/logging.h>
-#include "trainer/trainer.h"
-#include "utils/common.h"
-#ifndef GFLAGS_GFLAGS_H_
-  namespace gflags = google;
-#endif  // GFLAGS_GFLAGS_H_
+#include "singa.h"
 /**
  * \file main.cc is the main entry of SINGA, like the driver program for Hadoop.
  *
  * 1. Users register their own implemented classes, e.g., layer, updater, etc.
- * 2. Users prepare the google protobuf object for the model configuration and
- * the cluster configuration.
+ * 2. Users prepare the google protobuf object for the job configuration.
  * 3. Users call trainer to start the training.
  *
  * TODO
- * 1. Add helper functions for users to configure their model and cluster
- * easily, e.g., AddLayer(layer_type, source_layers, meta_data).
+ * 1. Add helper functions for users to configure their model easily,
+ * e.g., AddLayer(layer_type, source_layers, meta_data).
  */
 
-// TODO: re-organize ClusterProto, GlobalProto and ModelProto.
-DEFINE_int32(job, -1, "Unique job ID");
-DEFINE_bool(resume, false, "Resume from checkpoint");
-DEFINE_string(cluster, "examples/mnist/cluster.conf", "Cluster config file");
-DEFINE_string(model, "examples/mnist/conv.conf", "Model config file");
-DEFINE_string(global, "conf/singa.conf", "Global config file");
+DEFINE_int32(job, -1, "Unique job ID generated from singa-run.sh");
+DEFINE_bool(resume, false, "Resume from checkpoint passed at cmd line");
+DEFINE_string(workspace, "./workspace", "workspace passed at cmd line");
 
 /**
  * Register layers, and other customizable classes.
@@ -31,7 +21,8 @@ DEFINE_string(global, "conf/singa.conf", "Global config file");
  * If users want to use their own implemented classes, they should register
  * them here. Refer to the Worker::RegisterDefaultClasses()
  */
-void RegisterClasses(const singa::ModelProto& proto) {
+void RegisterClasses() {
+
 }
 
 
@@ -39,20 +30,15 @@ int main(int argc, char **argv) {
   google::InitGoogleLogging(argv[0]);
   gflags::ParseCommandLineFlags(&argc, &argv, true);
 
-  singa::ClusterProto cluster;
-  singa::ReadProtoFromTextFile(FLAGS_cluster.c_str(), &cluster);
-  singa::ModelProto model;
-  singa::ReadProtoFromTextFile(FLAGS_model.c_str(), &model);
-  singa::GlobalProto global;
-  singa::ReadProtoFromTextFile(FLAGS_global.c_str(), &global);
-  singa::SetupLog(global.log_dir(), model.name());
+  singa::JobProto jobConf;
+  std::string job_file = FLAGS_workspace + "/job.conf";
+  singa::ReadProtoFromTextFile(job_file.c_str(), &jobConf);
+  CHECK(jobConf.has_cluster());
+  CHECK(jobConf.has_model());
+  if (!jobConf.cluster().has_workspace())
+    jobConf.mutable_cluster()->set_workspace(FLAGS_workspace);
 
-  LOG(INFO) << "The cluster config is\n" << cluster.DebugString();
-  LOG(INFO) << "The model config is\n" << model.DebugString();
-  LOG(INFO) << "The global config is\n" << global.DebugString();
-
-  RegisterClasses(model);
-  singa::Trainer trainer;
-  trainer.Start(model, global, cluster, FLAGS_job, FLAGS_resume);
+  RegisterClasses();
+  singa::SubmitJob(FLAGS_job, FLAGS_resume, jobConf);
   return 0;
 }
