+++ src/neuralnet/layer.cc
@@ -1,8 +1,6 @@
 #include <glog/logging.h>
 #include <memory>
 #include <algorithm>
-#include <opencv2/highgui/highgui.hpp>
-#include <opencv2/imgproc/imgproc.hpp>
 #include "mshadow/tensor.h"
 #include "mshadow/cxxnet_op.h"
 #include "neuralnet/layer.h"
@@ -231,103 +229,6 @@ void LabelLayer::ParseRecords(Phase phase, const vector<Record>& records,
   CHECK_EQ(rid, blob->shape()[0]);
 }
 
-
-/*********************LMDBDataLayer**********************************/
-void LMDBDataLayer::ComputeFeature(Phase phase, Metric* perf){
-  if(random_skip_){
-    int nskip = rand() % random_skip_;
-    int n=0;
-    CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_,
-          &mdb_value_, MDB_FIRST), MDB_SUCCESS);
-    while (mdb_cursor_get(mdb_cursor_, &mdb_key_,
-          &mdb_value_, MDB_NEXT) == MDB_SUCCESS)
-      n++;
-    LOG(INFO)<<"Random Skip "<<nskip<<" records of total "<<n<<"records";
-    // We have reached the end. Restart from the first.
-    CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_,
-          &mdb_value_, MDB_FIRST), MDB_SUCCESS);
-    for(int i=0;i<nskip;i++){
-      if (mdb_cursor_get(mdb_cursor_, &mdb_key_,
-            &mdb_value_, MDB_NEXT) != MDB_SUCCESS) {
-        // We have reached the end. Restart from the first.
-        DLOG(INFO) << "Restarting data prefetching from start.";
-        CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_,
-              &mdb_value_, MDB_FIRST), MDB_SUCCESS);
-      }
-    }
-    random_skip_=0;
-  }
-  CaffeDatum datum;
-  for(auto& record: records_){
-    SingleLabelImageRecord* image=record.mutable_image();
-    CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_,
-          &mdb_value_, MDB_GET_CURRENT), MDB_SUCCESS);
-    datum.ParseFromArray(mdb_value_.mv_data, mdb_value_.mv_size);
-    ConvertCaffeDatumToRecord(datum, image);
-    if (mdb_cursor_get(mdb_cursor_, &mdb_key_,
-          &mdb_value_, MDB_NEXT) != MDB_SUCCESS) {
-      // We have reached the end. Restart from the first.
-      DLOG(INFO) << "Restarting data prefetching from start.";
-      CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_,
-            &mdb_value_, MDB_FIRST), MDB_SUCCESS);
-    }
-  }
-}
-
-void LMDBDataLayer::ConvertCaffeDatumToRecord(const CaffeDatum& datum,
-    SingleLabelImageRecord* record){
-  record->set_label(datum.label());
-  record->clear_shape();
-  if(datum.has_channels())
-    record->add_shape(datum.channels());
-  if(datum.has_height())
-    record->add_shape(datum.height());
-  if(datum.has_width())
-    record->add_shape(datum.width());
-  if(datum.has_data())
-    record->set_pixel(datum.data());
-  if(datum.float_data_size()){
-    record->clear_data();
-    for(float x: datum.float_data())
-      record->add_data(x);
-  }
-}
-
-void LMDBDataLayer::Setup(const LayerProto& proto, int npartitions) {
-  Layer::Setup(proto, npartitions);
-  CHECK_EQ(mdb_env_create(&mdb_env_), MDB_SUCCESS) << "mdb_env_create failed";
-  CHECK_EQ(mdb_env_set_mapsize(mdb_env_, 1099511627776), MDB_SUCCESS); // 1TB
-  CHECK_EQ(mdb_env_open(mdb_env_,
-        proto.lmdbdata_conf().path().c_str(),
-        MDB_RDONLY, 0664), MDB_SUCCESS) << "cannot open lmdb "
-    << proto.lmdbdata_conf().path();
-  CHECK_EQ(mdb_txn_begin(mdb_env_, NULL, MDB_RDONLY, &mdb_txn_), MDB_SUCCESS)
-    << "mdb_txn_begin failed";
-  CHECK_EQ(mdb_open(mdb_txn_, NULL, 0, &mdb_dbi_), MDB_SUCCESS)
-    << "mdb_open failed";
-  CHECK_EQ(mdb_cursor_open(mdb_txn_, mdb_dbi_, &mdb_cursor_), MDB_SUCCESS)
-    << "mdb_cursor_open failed";
-  LOG(INFO) << "Opening lmdb " << proto.lmdbdata_conf().path();
-  CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_, &mdb_value_, MDB_FIRST),
-      MDB_SUCCESS) << "mdb_cursor_get failed";
-
-  if (mdb_cursor_get(mdb_cursor_, &mdb_key_, &mdb_value_, MDB_NEXT)
-      != MDB_SUCCESS) {
-    CHECK_EQ(mdb_cursor_get(mdb_cursor_, &mdb_key_, &mdb_value_,
-          MDB_FIRST), MDB_SUCCESS);
-  }
-  CaffeDatum datum;
-  datum.ParseFromArray(mdb_value_.mv_data, mdb_value_.mv_size);
-  SingleLabelImageRecord* record=sample_.mutable_image();
-  ConvertCaffeDatumToRecord(datum, record);
-
-  batchsize_=batchsize();
-  if(partition_dim() == 0)
-    batchsize_ /= npartitions;
-  records_.resize(batchsize_);
-  random_skip_=proto.lmdbdata_conf().random_skip();
-}
-
 /***************** Implementation for LRNLayer *************************/
 void LRNLayer::Setup(const LayerProto& proto, int npartitions) {
   Layer::Setup(proto, npartitions);
@@ -377,63 +278,29 @@ void MnistLayer::ParseRecords(Phase phase,
   LOG_IF(ERROR, records.size()==0)<<"Empty records to parse";
   int ndim=records.at(0).image().shape_size();
   int inputsize =records.at(0).image().shape(ndim-1);
+  CHECK_EQ(inputsize, blob->shape()[1]);
 
   float* dptr=blob->mutable_cpu_data();
   for(const Record& record: records){
-    // copy from record to cv::Mat
-    cv::Mat input(inputsize, inputsize, CV_32FC1);
     const SingleLabelImageRecord& imagerecord=record.image();
-    if(imagerecord.pixel().size()){
+    if(imagerecord.pixel().size()) {
       string pixel=imagerecord.pixel();
-      for(int i=0,k=0;i<inputsize;i++)
-        for(int j=0;j<inputsize;j++)
+      for(int i = 0, k = 0; i < inputsize; i++) {
+        for(int j = 0; j < inputsize; j++) {
           // NOTE!!! must cast pixel to uint8_t then to float!!! waste a lot of
           // time to debug this
-          input.at<float>(i,j)=static_cast<float>(static_cast<uint8_t>(pixel[k++]));
-    }else{
-      for(int i=0,k=0;i<inputsize;i++)
-        for(int j=0;j<inputsize;j++)
-          input.at<float>(i,j)=imagerecord.data(k++);
-    }
-    int size=blob->shape()[1];
-    /*
-    cv::Mat resizeMat=input;
-    // affine transform, scaling, rotation and shearing
-    if(gamma_){
-      float r1=rand_real()*2-1;
-      float r2=rand_real()*2-1;
-      int h=static_cast<int>(inputsize*(1.+r1*gamma_/100.0));
-      int w=static_cast<int>(inputsize*(1.+r2*gamma_/100.0));
-      cv::resize(input, resizeMat, cv::Size(h,w));
-    }
-    cv::Mat betaMat=resizeMat;
-    cv::Mat warpmat(2,3, CV_32FC1);
-    warpmat.at<float>(0,0)=1.0;
-    warpmat.at<float>(0,1)=0.0;
-    warpmat.at<float>(0,2)=0.0;
-    warpmat.at<float>(1,0)=0.0;
-    warpmat.at<float>(1,1)=1.0;
-    warpmat.at<float>(1,2)=0.0;
-
-    if(beta_){
-      float r=rand_real()*2-1;
-      if(rand() % 2){ // rotation
-        cv::Point center(resizeMat.rows/2, resizeMat.cols/2);
-        warpmat=cv::getRotationMatrix2D(center, r*beta_, 1.0);
-      }else{
-        //shearing
-        warpmat.at<float>(0,1)=r*beta_/90;
-        if(imagerecord.label()==1 ||imagerecord.label()==7)
-          warpmat.at<float>(0,1)/=2.0;
+          float x =  static_cast<float>(static_cast<uint8_t>(pixel[k++]));
+          x = x / norm_a_-norm_b_;
+          *dptr = x;
+          dptr++;
+        }
       }
-    }
-    cv::warpAffine(resizeMat, betaMat, warpmat, cv::Size(size, size));
-    */
-
-    for(int i=0;i<size;i++){
-      for(int j=0;j<size;j++){
-        *dptr=input.at<float>(i,j)/norm_a_-norm_b_;
-        dptr++;
+    } else {
+      for(int i = 0, k = 0; i < inputsize; i++) {
+        for(int j = 0; j < inputsize; j++) {
+          *dptr = imagerecord.data(k++) / norm_a_ - norm_b_;
+          dptr++;
+        }
       }
     }
   }
